examples/python-deploy-contract/requirements.txt
Arpeggio==2.0.0
astroid==2.15.0
attrs==22.1.0
autopep8==1.6.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
dill==0.3.6
docstring-to-markdown==0.11
flake8==6.0.0
google-auth==2.19.0
grpcio==1.56.2
grpcio-tools==1.56.2
idna==3.4
isort==5.12.0
jedi==0.18.2
kubernetes==26.1.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
parver==0.4
platformdirs==3.1.1
pluggy==1.0.0
protobuf==4.21.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydocstyle==6.2.3
pyflakes==3.0.1
pylint==2.17.0
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rope==1.7.0
rsa==4.9
semver==2.13.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tomlkit==0.11.6
ujson==5.7.0
urllib3==1.26.13
websocket-client==1.5.2
whatthepatch==1.0.4
wrapt==1.15.0
yapf==0.32.0
